* looks as if observer.js will be the one!!
using proxies doesn't seem to buy me anything and would be cumbersome since we'd have to pass the proxy back to the caller to use. observer.js just adds getters and setters to the original object. If there were previously defined getters / setters, it will call them as well to preserve their effects!